Added login activity and the app is now connected to a global database. Also added a userId field to the local database, and added functionality for syncing hikes from local DB to global.

+++ MapApp/app/src/main/java/msh/frida/mapapp/HikeSummaryActivity.java
@@ -25,7 +25,7 @@ public class HikeSummaryActivity extends AppCompatActivity {
         setContentView(R.layout.activity_hike_summary);
 
         Bundle extras = getIntent().getExtras();
-        int hikeId = extras.getInt("hikeId");
+        int hikeId = (int) extras.getLong("hikeId");
         //double distanceWalked = extras.getDouble("distanceWalked");
 
         DatabaseHandler db = new DatabaseHandler(this);
